minor bug fix + per field validation
